requirements.txt
flask
gevent
pyzmq
boto
requests==1.2.0 # need 1.2.0 since docker-py...
redis
sqlalchemy
fabric
etcd-py
docker-py